Move Phoenix.HTML.Tag.tag/content_tag to Phoenix.HTML
