+++ test/phoenix_html/tag_test.exs
@@ -5,97 +5,6 @@ defmodule Phoenix.HTML.TagTest do
   import Phoenix.HTML.Tag, except: [attributes_escape: 1]
   doctest Phoenix.HTML.Tag
 
-  test "tag" do
-    assert tag(:br) |> safe_to_string() == ~s(<br>)
-
-    assert tag(:input, name: ~s("<3")) |> safe_to_string() == ~s(<input name="&quot;&lt;3&quot;">)
-    assert tag(:input, name: raw("<3")) |> safe_to_string() == ~s(<input name="<3">)
-    assert tag(:input, name: ["foo", raw("b<r")]) |> safe_to_string() == ~s(<input name="foob<r">)
-    assert tag(:input, name: :hello) |> safe_to_string() == ~s(<input name="hello">)
-
-    assert tag(:input, type: "text", name: "user_id") |> safe_to_string() ==
-             ~s(<input name="user_id" type="text">)
-
-    assert tag(:input, data: [toggle: "dropdown"]) |> safe_to_string() ==
-             ~s(<input data-toggle="dropdown">)
-
-    assert tag(:input, my_attr: "blah") |> safe_to_string() == ~s(<input my-attr="blah">)
-
-    assert tag(:input, [{"my_<_attr", "blah"}]) |> safe_to_string() ==
-             ~s(<input my_&lt;_attr="blah">)
-
-    assert tag(:input, [{{:safe, "my_<_attr"}, "blah"}]) |> safe_to_string() ==
-             ~s(<input my_<_attr="blah">)
-
-    assert tag(:input, data: [my_attr: "blah"]) |> safe_to_string() ==
-             ~s(<input data-my-attr="blah">)
-
-    assert tag(:input, data: [toggle: [attr: "blah", target: "#parent"]]) |> safe_to_string() ==
-             ~s(<input data-toggle-attr="blah" data-toggle-target="#parent">)
-
-    assert tag(:audio, autoplay: "autoplay") |> safe_to_string() ==
-             ~s(<audio autoplay="autoplay">)
-
-    assert tag(:audio, autoplay: true) |> safe_to_string() == ~s(<audio autoplay>)
-    assert tag(:audio, autoplay: false) |> safe_to_string() == ~s(<audio>)
-    assert tag(:audio, autoplay: nil) |> safe_to_string() == ~s(<audio>)
-  end
-
-  test "content_tag" do
-    assert content_tag(:p, "Hello") |> safe_to_string() == "<p>Hello</p>"
-
-    assert content_tag(:p, "Hello", class: "dark") |> safe_to_string() ==
-             "<p class=\"dark\">Hello</p>"
-
-    assert content_tag(:p, [class: "dark"], do: "Hello") |> safe_to_string() ==
-             "<p class=\"dark\">Hello</p>"
-
-    assert content_tag(:p, "<Hello>") |> safe_to_string() == "<p>&lt;Hello&gt;</p>"
-
-    assert content_tag(:p, 13) |> safe_to_string() == "<p>13</p>"
-
-    assert content_tag(:p, [class: "dark"], do: "<Hello>") |> safe_to_string() ==
-             "<p class=\"dark\">&lt;Hello&gt;</p>"
-
-    assert content_tag(:p, raw("<Hello>")) |> safe_to_string() == "<p><Hello></p>"
-
-    assert content_tag(:p, [class: "dark"], do: raw("<Hello>")) |> safe_to_string() ==
-             "<p class=\"dark\"><Hello></p>"
-
-    content =
-      content_tag :form, action: "/users", data: [remote: true] do
-        tag(:input, name: "user[name]")
-      end
-
-    assert safe_to_string(content) ==
-             ~s(<form action="/users" data-remote="true">) <> ~s(<input name="user[name]"></form>)
-
-    assert content_tag(:p, do: "Hello") |> safe_to_string() == "<p>Hello</p>"
-
-    content =
-      content_tag :ul do
-        content_tag :li do
-          "Hello"
-        end
-      end
-
-    assert safe_to_string(content) == "<ul><li>Hello</li></ul>"
-
-    assert content_tag(:p, ["hello", ?\s, "world"]) |> safe_to_string() == "<p>hello world</p>"
-
-    assert content_tag(:div, [autoplay: "autoplay"], do: "") |> safe_to_string() ==
-             ~s(<div autoplay="autoplay"></div>)
-
-    assert content_tag(:div, [autoplay: true], do: "") |> safe_to_string() ==
-             ~s(<div autoplay></div>)
-
-    assert content_tag(:div, [autoplay: false], do: "") |> safe_to_string() == ~s(<div></div>)
-
-    assert content_tag(:div, [autoplay: nil], do: "") |> safe_to_string() == ~s(<div></div>)
-
-    assert content_tag("custom-tag", "Hi") |> safe_to_string() == ~s(<custom-tag>Hi</custom-tag>)
-  end
-
   test "img_tag" do
     assert img_tag("user.png") |> safe_to_string() == ~s(<img src="user.png">)
